RFC1078 TCPMUX help
help is reserved keyword according to rfc1078 and must list TCPMUX
services

This is an implementation of the help command for the tcpmux-server
builtin service included in xinetd.

The patch also introduces new flag TCPMUXHELP which is required for
the tcpmux service to be published via the help command.  TCPMUX and
TCPMUXPLUS services which do not have the TCPMUXHELP flag are not listed
in order to preserve the existing behavior and to support providing
unpublished TCPMUX services.