new feature: support for rfc1078 help command
"help" is reserved keyword according to rfc1078 and must list TCPMUX
services.  If the help service is requested from the TCPMUX server,
any services listed must be accessible via TCPMUX.

The patch also introduces new flag PUBLISH which is required for
the tcpmux service to be published via the help command. TCPMUX and
TCPMUXPLUS services which do not have the PUBLISH flag are not listed
in order to preserve the existing behavior and to support providing
unpublished TCPMUX services.