xinetd should be able to listen on IPv6 even in -inetd_compat mode
xinetd does not bind to IPv6 addresses (and does not seems to have an
option to do so) when used in -inetd_compat mode.  As current inetd's
are IPv6-aware, this is a problem: this means xinetd
cannot be used as a drop-in inetd replacement.

The attached patch is a suggestion: it adds a -inetd_ipv6 global
option that, if used, causes inetd-compatibility lines to have an
implicit "IPv6" option.  Perhaps this is not the best solution, but
there should definitely be a way to get inetd.conf to be read in
IPv6-aware mode

Signed-off-by: David Madore <david@pleiades.stars>
Signed-off-by: Maxin B. John <maxin.john@intel.com>