Prevent deadlocks in the signal handler bsc#726737
A signal can interrupt xinetd when it's printing / logging.

Currently, the generic signal handler tries to print something as
well (the "Unexpected signal" message), but to do that it needs a
lock that xinetd already holds, deadlocking.

We really can't print anything in the signal handler if the process
is supposed to continue running.

This patch is not perfect but it avoids the most common cases by
no longer catching SIGCONT and no longer printing the "Unexpected
signal" message in the generic signal handler.